+++ lib/rh-admin-utils/WpscHtaccessHelper.php
@@ -9,7 +9,7 @@ namespace RH\AdminUtils;
 if (!defined('ABSPATH')) exit; // Exit if accessed directly
 
 /**
- * A helper to modify default WP Super Cache behavior
+ * A helper to modify the WP Super Cache cache/.htaccess
  */
 class WpscHtaccessHelper
 {
@@ -22,7 +22,53 @@ class WpscHtaccessHelper
      */
     public static function init()
     {
-        add_action('wp_cache_cleared', [__CLASS__, 'modify_cache_dir_htaccess']);
+        add_action('wp_cache_cleared', [__CLASS__, 'add_custom_directives_to_htaccess']);
+        add_action('wp_cache_cleared', [__CLASS__, 'modify_supercache_directives']);
+        add_action(
+            'admin_init',
+            [__CLASS__, 'modify_htaccess_on_super_cache_options_update'],
+            11 // execute just after wp_cache_manager_updates has completed its thing
+        );
+    }
+
+    /**
+     * Modifies the .htaccess every time the Super Cache
+     * Settings > "Update Status" button is clicked
+     *
+     * @return void
+     */
+    public static function modify_htaccess_on_super_cache_options_update() {
+        $action = $_POST['action'] ?? null;
+        $is_super_cache_update = $action === 'scupdates';
+        if (!$is_super_cache_update) return;
+        self::add_custom_directives_to_htaccess();
+        self::modify_supercache_directives();
+    }
+
+    /**
+     * Get the .htaccess path
+     *
+     * @return string
+     */
+    private static function get_htaccess_path(): string {
+        return WP_CONTENT_DIR . '/cache/.htaccess';
+    }
+
+    /**
+     * Checks if the wp-content/cache/.htaccess exists
+     */
+    private static function htaccess_exists(): bool
+    {
+        return file_exists(self::get_htaccess_path());
+    }
+
+    /**
+     * Make sure the required api functions are available
+     */
+    private static function require_wp_api_functions(): void
+    {
+        if (function_exists('insert_with_markers')) return;
+        require_once(ABSPATH . 'wp-admin/includes/misc.php');
     }
 
     /**
@@ -30,16 +76,10 @@ class WpscHtaccessHelper
      *
      * @return void
      */
-    public static function modify_cache_dir_htaccess(): void
+    public static function add_custom_directives_to_htaccess(): void
     {
-        // Check if the wp-content/cache/.htaccess exists
-        $htaccess_path = WP_CONTENT_DIR . '/cache/.htaccess';
-        if (!file_exists($htaccess_path)) return;
-
-        // Make sure `insert_with_markers` exists
-        if (!function_exists('insert_with_markers')) {
-            require_once(ABSPATH . 'wp-admin/includes/misc.php');
-        }
+        if (!self::htaccess_exists()) return;
+        self::require_wp_api_functions();
 
         // Our custom directives
         $custom_directives = [
@@ -50,10 +90,60 @@ class WpscHtaccessHelper
 
         // Update the .htaccess file
         insert_with_markers(
-            $htaccess_path,
+            self::get_htaccess_path(),
             'rh-admin-utils',
             $custom_directives
         );
     }
 
+    /**
+     * Modify the supercache directives
+     *
+     * @return void
+     */
+    public static function modify_supercache_directives(): void
+    {
+
+        if (!self::htaccess_exists()) return;
+        self::require_wp_api_functions();
+
+        // Extract the supercache directives
+        $directives = extract_from_markers(self::get_htaccess_path(), 'supercache');
+        // Trim each directive, to make it easier to handle them
+        $directives = array_map('trim', $directives);
+
+        $new_directives = [];
+
+        // Make the two dirctives I want to customize filterable
+        $cache_control_header = apply_filters(
+            "rhau/wpsc/cache-control-header",
+            "max-age=3, private"
+        );
+        $expires_by_type = apply_filters(
+            "rhau/wpsc/expires-by-type",
+            "access plus 3 seconds"
+        );
+
+        foreach ($directives as $directive) {
+            // Modifiy the Cache-Control header directive
+            if (str_starts_with($directive, "Header set Cache-Control")) {
+                $new_directives[] = "Header set Cache-Control '$cache_control_header'";
+                continue;
+            }
+            // Modify the ExpiresByType directive
+            if (str_starts_with($directive, "ExpiresByType text/html")) {
+                $new_directives[] = "ExpiresByType text/html '$expires_by_type'";
+                continue;
+            }
+            // Add all other directives without modification
+            $new_directives[] = $directive;
+        }
+
+        // Update the .htaccess file
+        insert_with_markers(
+            self::get_htaccess_path(),
+            'supercache',
+            $new_directives
+        );
+    }
 }